ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
transformers==4.41.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
hypothesis==6.100.0
pytest==8.2.0
httpx==0.27.0
python-dotenv==1.0.1
